# Let memdb cache last traversed node