--- a/internal/unionstore/union_store.go
+++ b/internal/unionstore/union_store.go
@@ -244,10 +244,9 @@ type MemBuffer interface {
 }
 
 type Metrics struct {
-	WaitDuration     time.Duration
-	MemDBHitCount    uint64
-	MemDBMissCount   uint64
-	TraverseDuration time.Duration
+	WaitDuration   time.Duration
+	MemDBHitCount  uint64
+	MemDBMissCount uint64
 }
 
 var (